+++ src/bewegung/linalg/_single.py
@@ -29,8 +29,10 @@ specific language governing rights and limitations under the License.
 # +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
 from abc import ABC, abstractmethod
+from typing import Union
 
 from ..lib import typechecked
+from ._abc import MetaDict
 
 # +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 # CLASS
@@ -42,8 +44,20 @@ class Vector(ABC):
     Abstract base class for all vector types.
 
     Not intended to be instantiated.
+
+    Args:
+        meta : A dict holding arbitrary metadata
     """
 
     @abstractmethod
-    def __init__(self): # pragma: no cover
-        pass
+    def __init__(self, meta: Union[MetaDict, None] = None):
+
+        self._meta = {} if meta is None else dict(meta)
+
+    @property
+    def meta(self) -> MetaDict:
+        """
+        meta data dict
+        """
+
+        return self._meta